+++ lexforge20/src/main/java/dev/nolij/nolijium/mixin/lexforge20/ChatComponentMixin.java
@@ -9,7 +9,10 @@ import org.spongepowered.asm.mixin.injection.ModifyConstant;
 @Mixin(ChatComponent.class)
 public class ChatComponentMixin {
 	
-	@ModifyConstant(method = "addMessage(Lnet/minecraft/network/chat/Component;Lnet/minecraft/network/chat/MessageSignature;ILnet/minecraft/client/GuiMessageTag;Z)V", constant = @Constant(intValue = 100))
+	@ModifyConstant(
+		method = "addMessage(Lnet/minecraft/network/chat/Component;Lnet/minecraft/network/chat/MessageSignature;ILnet/minecraft/client/GuiMessageTag;Z)V", 
+		constant = @Constant(intValue = 100)
+	)
 	public int nolijium$MAX_CHAT_HISTORY(int constant) {
 		return Nolijium.config.maxChatHistory > 0 
 		       ? Nolijium.config.maxChatHistory